fix: renames init to setup
